+++ fuzz/fuzz_targets/roundtrip.rs
@@ -1,0 +1,9 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+use shortguid::ShortGuid;
+
+fuzz_target!(|bytes: [u8; 16]| {
+    let id = ShortGuid::from_bytes_ref(&bytes);
+    assert_eq!(&bytes, id.as_bytes());
+});