+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- `from_bytes_ref`, `as_uuid` and `is_empty` are now `const`.
+
+### Internal
+
+- Added fuzzing targets.
+
 ## 0.1.0 - 2023-04-07
 
 ### Added